requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
dropbox==11.36.2
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pymysql==1.1.0
cryptography==41.0.7

